[ALPS03194742] Display: fix the Coverity issue.
[Detail] fix the coverity issue.

MTK-Commit-Id: fe1e9f06cfe35062224f1d1b755eaee72a3c0502

CR-Id: ALPS03194742
Feature: Display Driver (DSI Interface)

+++ drivers/misc/mediatek/video/mt6763/videox/disp_lowpower.c
@@ -1188,7 +1188,7 @@ unsigned int time_to_line(unsigned int ms, unsigned int width)
 	unsigned int line = 0;
 
 	tline_us = get_us_perline(width);
-	time_us = ms * 1000 * LINE_ACCURACY;
+	time_us = (unsigned long long)ms * 1000 * LINE_ACCURACY;
 #if defined(__LP64__) || defined(_LP64)
 	line = time_us / tline_us;
 #else
